+++ stohastic/stat.h
@@ -3,11 +3,11 @@
 using namespace std;
 
 void set_g_rnd_seed(unsigned int seed);
-double sigma(double *x, int n);
-double max_arr(double *x, int n);
-double min_arr(double *x, int n);
+long double sigma(long double *x, int n);
+long double max_arr(long double *x, int n);
+long double min_arr(long double *x, int n);
 inline unsigned int rnd_i();
 inline unsigned int rnd_u();
-inline double rnd();
-double rnd_g();
+inline long double rnd();
+long double rnd_g();
 #endif
